Loader: split load_test() method into a common module
This introduces a new core module to hold a function that is required
both by the current runner and by the avocado-instrumented nrunner.
