+++ avocado/core/runners/avocado_instrumented.py
@@ -3,7 +3,7 @@ import tempfile
 import time
 import traceback
 
-from .. import loader, nrunner
+from .. import miniloader, nrunner
 from ..test import TestID
 from ..tree import TreeNode, TreeNodeEnvOnly
 from ..varianter import is_empty_variant
@@ -75,7 +75,7 @@ class AvocadoInstrumentedTestRunner(nrunner.BaseRunner):
                              }]
 
             messages.start_logging(runnable.config, queue)
-            instance = loader.loader.load_test(test_factory)
+            instance = miniloader.load_test(test_factory)
             early_state = instance.get_state()
             early_state['type'] = "early_state"
             queue.put(early_state)
